Random memory
The temple as well as the temple analyzer have been updated to create a random memory value for each prosses. Along with this sense my multithreading feature created new programs it would now make it so that child prosses would have a fraction of the parent based on the number of operations.