paging and bug fixes
now in the OS is the option between cautious and nonconscious memory upon initialization. In order to easy switch between these two I made a costume class to take care of all page related functionality. This includes creation, termination, and victim selection. Along with a new way in which the cpu can view the processes it is choosing to run. Along side this there was also bugs in the forking algorithm due to memory sharing among child and parent prosses.