bugged multithreading patch
the simulation has worked for me on some inputs with more then 1 cpu and typically with programs in the hundreds. I suspect the main problem is a reader writer problem that with this configuration is hard to fix without changing CPU.java into several smaller programs. It should produce it's own scheduler class and the pcb store all random simulated prosses from the initialization.